+++ pkg/state/errors.go
@@ -7,8 +7,6 @@ package state
 import (
 	"errors"
 	"fmt"
-
-	"github.com/voedger/voedger/pkg/appdef"
 )
 
 var ErrNotSupported = errors.New("not supported")
@@ -30,7 +28,6 @@ var errWorkspaceDescriptorNotFound = errors.New("WorkspaceDescriptor not found i
 var errDescriptorForUndefinedWorkspace = errors.New("workspace descriptor for undefined workspace")
 var errCommandNotSpecified = errors.New("command not specified")
 var errBlobIDNotSpecified = errors.New("blob ID not specified")
-var ErrQNameIsNotDefinedInWorkspace = errors.New("qname is not defined in workspace")
 
 func errUnexpectedType(actual interface{}) error {
 	return fmt.Errorf("unexpected type: %v", actual)
@@ -38,8 +35,4 @@ func errUnexpectedType(actual interface{}) error {
 
 func errIndexOutOfBounds(index int) error {
 	return fmt.Errorf("index out of bounds: %d", index)
-}
-
-func typeIsNotDefinedInWorkspaceWithDescriptor(typ, ws appdef.QName) error {
-	return fmt.Errorf("%s %w %s", typ.String(), ErrQNameIsNotDefinedInWorkspace, ws.String())
 }
