+++ pkg/sys/storages/impl_commandcontext_storage_test.go
@@ -2,7 +2,7 @@
  * Copyright (c) 2023-present unTill Pro, Ltd.
  */
 
-package state
+package storages
 
 import (
 	"context"
