+++ pkg/sys/storages/impl_event_storage_test.go
@@ -2,7 +2,7 @@
  * Copyright (c) 2022-present unTill Pro, Ltd.
  */
 
-package state
+package storages
 
 import (
 	"context"
