+++ pkg/sys/storages/impl_subject_storage_test.go
@@ -2,7 +2,7 @@
  * Copyright (c) 2020-present unTill Pro, Ltd.
  */
 
-package state
+package storages
 
 import (
 	"context"
